requirements.txt
discord.py==1.7.3
mysql-connector-python==8.0.22
requests==2.28
